requirements.txt
Python 3.9
Django <=3.2
Djangorestframework
Django-filter
Celery
Redis
django-celery-beat
djoser